Improve SEO capturing
