+++ kryptone/contrib/crawlers/seo.py
@@ -1,14 +1,15 @@
+import asyncio
+
 from kryptone import logger
 from kryptone.base import SiteCrawler
 from kryptone.mixins import SEOMixin
 from kryptone.utils import file_readers
-import asyncio
 
 
 class SEOCrawler(SiteCrawler, SEOMixin):
     """A crawler specialized for running
     SEO tasks on a given website"""
-    
+
     def resume(self, **kwargs):
         data = file_readers.read_json_document('cache.json')
 
@@ -19,14 +20,16 @@ class SEOCrawler(SiteCrawler, SEOMixin):
         self.urls_to_visit = set(valid_urls)
         self.visited_urls = set(data['visited_urls'])
 
-        previous_seen_urls = file_readers.read_csv_document('seen_urls.csv', flatten=True)
+        previous_seen_urls = file_readers.read_csv_document(
+            'seen_urls.csv', flatten=True)
         self.list_of_seen_urls = set(previous_seen_urls)
 
         self.page_audits = file_readers.read_json_document('audit.json')
-        self.text_by_page = file_readers.read_json_document('text_by_pages.json')
+        self.text_by_page = file_readers.read_json_document(
+            'text_by_pages.json')
 
         self.start(**kwargs)
-    
+
     def run_actions(self, current_url, **kwargs):
         self.audit_page(current_url)
         vocabulary = self.global_audit()
@@ -46,7 +49,7 @@ class SEOCrawler(SiteCrawler, SEOMixin):
             # Save the website's text
             website_text = ' '.join(self.fitted_page_documents)
             file_readers.write_text_document(
-                'website_text.txt', 
+                'website_text.txt',
                 website_text
             )
 
@@ -60,12 +63,12 @@ class SEOCrawler(SiteCrawler, SEOMixin):
             task2 = asyncio.create_task(write_vocabulary())
             task3 = asyncio.create_task(write_website_text())
             tasks = [task1, task2, task3]
-            
+
             for coroutine in asyncio.as_completed(tasks):
                 await coroutine
 
         asyncio.run(main())
-        
+
         # db_signal.send(
         #     self,
         #     page_audit=self.page_audits,
